# Fix: forced no-shadow to check all scopes (fixes #2294)