--- a/lib/rules/no-shadow.js
+++ b/lib/rules/no-shadow.js
@@ -49,11 +49,11 @@ module.exports = function(context) {
 
     /**
      * Checks the current context for shadowed variables.
+     * @param {Scope} scope - Fixme
      * @returns {void}
      */
-    function checkForShadows() {
-        var scope = context.getScope(),
-            variables = scope.variables;
+    function checkForShadows(scope) {
+        var variables = scope.variables;
 
         // iterate through the array of variables and find duplicates with the upper scope
         var upper = scope.upper;
@@ -64,9 +64,17 @@ module.exports = function(context) {
     }
 
     return {
-        "FunctionDeclaration": checkForShadows,
-        "FunctionExpression": checkForShadows,
-        "ArrowFunctionExpression": checkForShadows
+        "Program:exit": function () {
+            var globalScope = context.getScope(),
+                stack = globalScope.childScopes.slice(),
+                scope;
+
+            while (stack.length) {
+                scope = stack.pop();
+                stack.push.apply(stack, scope.childScopes);
+                checkForShadows(scope);
+            }
+        }
     };
 
 };