Instead of listing all the nodes that could potentially create a scope in ES6, I just built a stack to traverse the scopes tree top-down.

Performance looks OK in my machine:

* Before:

    ```
    CPU Speed is 800 with multiplier 7500000
    Performance Run #1:  1276.091608ms
    Performance Run #2:  1313.046001ms
    Performance Run #3:  1282.84558ms
    Performance Run #4:  1280.500036ms
    Performance Run #5:  1308.158333ms
    Performance budget ok:  1282.84558ms (limit: 9375ms)
    ```
* After:

    ```
    CPU Speed is 800 with multiplier 7500000
    Performance Run #1:  1264.320176ms
    Performance Run #2:  1256.536445ms
    Performance Run #3:  1258.839234ms
    Performance Run #4:  1276.002489ms
    Performance Run #5:  1273.202285ms
    Performance budget ok:  1264.320176ms (limit: 9375ms)
    ```